@@ -1,178 +1,296 @@
-# Copyright (c) 2016-2017, Neil Booth
+# Copyright (c) 2016-2021, Neil Booth
 # Copyright (c) 2017, the ElectrumX authors
 #
 # All rights reserved.
 #
-# See the file "LICENCE" for information about the copyright
-# and warranty status of this software.
+# This file is licensed under the Open BSV License version 3, see LICENCE for details.
 
 '''Block prefetcher and chain processor.'''
 
 
 import asyncio
-import time
+import os
+import re
+from asyncio import sleep
+from datetime import datetime
+from struct import error as struct_error
 
-from aiorpcx import TaskGroup, run_in_thread, CancelledError
+from aiorpcx import CancelledError, run_in_thread, spawn
 
 import electrumx
-from electrumx.server.daemon import DaemonError
 from electrumx.lib.hash import hash_to_hex_str, HASHX_LEN
 from electrumx.lib.script import is_unspendable_legacy, is_unspendable_genesis
+from electrumx.lib.tx import Deserializer
 from electrumx.lib.util import (
-    chunks, class_logger, pack_le_uint32, pack_le_uint64, unpack_le_uint64
+    class_logger, pack_le_uint32, pack_le_uint64, unpack_le_uint64, open_file, unpack_le_uint32,
 )
 from electrumx.server.db import FlushData
 
 
-class Prefetcher(object):
-    '''Prefetches blocks (in the forward direction only).'''
+logger = class_logger(__name__, 'BlockProcessor')
 
-    def __init__(self, daemon, coin, blocks_event):
-        self.logger = class_logger(__name__, self.__class__.__name__)
-        self.daemon = daemon
-        self.coin = coin
-        self.blocks_event = blocks_event
-        self.blocks = []
-        self.caught_up = False
-        # Access to fetched_height should be protected by the semaphore
-        self.fetched_height = None
-        self.semaphore = asyncio.Semaphore()
-        self.refill_event = asyncio.Event()
-        # The prefetched block cache size.  The min cache size has
-        # little effect on sync time.
-        self.cache_size = 0
-        self.min_cache_size = 10 * 1024 * 1024
-        # This makes the first fetch be 10 blocks
-        self.ave_size = self.min_cache_size // 10
-        self.polling_delay = 5
 
-    async def main_loop(self, bp_height):
-        '''Loop forever polling for more blocks.'''
-        await self.reset_height(bp_height)
+class OnDiskBlock:
+
+    path = 'meta/blocks'
+    legacy_del_regex = re.compile('block[0-9]{1,7}$')
+    block_regex = re.compile('([0-9]{1,7})-([0-9a-f]{64})$')
+    chunk_size = 25_000_000
+    # On-disk blocks. hex_hash->(height, size) pair
+    blocks = {}
+    # Map from hex hash to prefetch task
+    tasks = {}
+    # If set it logs the next time a block is processed
+    log_block = False
+    daemon = None
+    state = None
+
+    def __init__(self, hex_hash, height, size):
+        self.hex_hash = hex_hash
+        self.height = height
+        self.size = size
+        self.block_file = None
+        self.header = None
+
+    @classmethod
+    def filename(cls, hex_hash, height):
+        return os.path.join(cls.path, f'{height:d}-{hex_hash}')
+
+    def __enter__(self):
+        self.block_file = open_file(self.filename(self.hex_hash, self.height))
+        self.header = self._read(80)
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.block_file.close()
+
+    def _read(self, size):
+        result = self.block_file.read(size)
+        if not result:
+            raise RuntimeError(f'truncated block file for block {self.hex_hash} '
+                               f'height {self.height:,d}')
+        return result
+
+    def _read_at_pos(self, pos, size):
+        self.block_file.seek(pos, os.SEEK_SET)
+        result = self.block_file.read(size)
+        if len(result) != size:
+            raise RuntimeError(f'truncated block file for block {self.hex_hash} '
+                               f'height {self.height:,d}')
+        return result
+
+    def date_str(self):
+        timestamp, = unpack_le_uint32(self.header[68:72])
+        return datetime.utcfromtimestamp(timestamp).strftime('%Y-%m-%d %H:%M:%S')
+
+    def iter_txs(self):
+        # Asynchronous generator of (tx, tx_hash) pairs
+        raw = self._read(self.chunk_size)
+        deserializer = Deserializer(raw)
+        tx_count = deserializer.read_varint()
+
+        if self.log_block:
+            logger.info(f'height {self.height:,d} of {self.daemon.cached_height():,d} '
+                        f'{self.hex_hash} {self.date_str()} '
+                        f'{self.size / 1_000_000_000:.3f}GB {tx_count:,d} txs '
+                        f'chain {self.state.chain_size // 1_000_000_000:,d}GB')
+            OnDiskBlock.log_block = False
+
+        count = 0
         while True:
+            read = deserializer.read_tx_and_hash
             try:
-                # Sleep a while if there is nothing to prefetch
-                await self.refill_event.wait()
-                if not await self._prefetch_blocks():
-                    await asyncio.sleep(self.polling_delay)
-            except DaemonError as e:
-                self.logger.info(f'ignoring daemon error: {e}')
-            except asyncio.CancelledError as e:
-                self.logger.info(f'cancelled; prefetcher stopping {e}')
-                raise
-            except Exception:
-                self.logger.exception(f'ignoring unexpected exception')
+                while True:
+                    cursor = deserializer.cursor
+                    yield read()
+                    count += 1
+            except (AssertionError, IndexError, struct_error):
+                pass
 
-    def get_prefetched_blocks(self):
-        '''Called by block processor when it is processing queued blocks.'''
-        blocks = self.blocks
-        self.blocks = []
-        self.cache_size = 0
-        self.refill_event.set()
-        return blocks
+            if count == tx_count:
+                return
+            raw = raw[cursor:] + self._read(self.chunk_size)
+            deserializer = Deserializer(raw)
 
-    async def reset_height(self, height):
-        '''Reset to prefetch blocks from the block processor's height.
+    def _chunk_offsets(self):
+        '''Iterate the transactions forwards to find their boundaries.'''
+        base_offset = self.block_file.tell()
+        assert base_offset == 80
+        raw = self._read(self.chunk_size)
+        deserializer = Deserializer(raw)
+        tx_count = deserializer.read_varint()
+        logger.info(f'backing up block {self.hex_hash} height {self.height:,d} '
+                    f'tx_count {tx_count:,d}')
+        offsets = [base_offset + deserializer.cursor]
 
-        Used in blockchain reorganisations.  This coroutine can be
-        called asynchronously to the _prefetch_blocks coroutine so we
-        must synchronize with a semaphore.
-        '''
-        async with self.semaphore:
-            self.blocks.clear()
-            self.cache_size = 0
-            self.fetched_height = height
-            self.refill_event.set()
+        while True:
+            read = deserializer.read_tx
+            count = 0
+            try:
+                while True:
+                    cursor = deserializer.cursor
+                    read()
+                    count += 1
+            except (AssertionError, IndexError, struct_error):
+                pass
 
-        daemon_height = await self.daemon.height()
-        behind = daemon_height - height
-        if behind > 0:
-            self.logger.info('catching up to daemon height {:,d} '
-                             '({:,d} blocks behind)'
-                             .format(daemon_height, behind))
-        else:
-            self.logger.info('caught up to daemon height {:,d}'
-                             .format(daemon_height))
+            if count:
+                offsets.append(base_offset + cursor)
+                base_offset += cursor
+            tx_count -= count
+            if tx_count == 0:
+                return offsets
+            raw = raw[cursor:] + self._read(self.chunk_size)
+            deserializer = Deserializer(raw)
 
-    async def _prefetch_blocks(self):
-        '''Prefetch some blocks and put them on the queue.
+    def iter_txs_reversed(self):
+        # Iterate the block transactions in reverse order.  We need to iterate the
+        # transactions forwards first to find their boundaries.
+        offsets = self._chunk_offsets()
+        for n in reversed(range(len(offsets) - 1)):
+            start = offsets[n]
+            size = offsets[n + 1] - start
+            deserializer = Deserializer(self._read_at_pos(start, size))
+            pairs = []
+            while deserializer.cursor < size:
+                pairs.append(deserializer.read_tx_and_hash())
+            for item in reversed(pairs):
+                yield item
 
-        Repeats until the queue is full or caught up.
-        '''
-        daemon = self.daemon
-        daemon_height = await daemon.height()
-        async with self.semaphore:
-            while self.cache_size < self.min_cache_size:
-                first = self.fetched_height + 1
-                # Try and catch up all blocks but limit to room in cache.
-                cache_room = max(self.min_cache_size // self.ave_size, 1)
-                count = min(daemon_height - self.fetched_height, cache_room)
-                # Don't make too large a request
-                count = min(self.coin.max_fetch_blocks(first), max(count, 0))
-                if not count:
-                    self.caught_up = True
-                    return False
+    @classmethod
+    async def delete_stale(cls, items, log):
+        def delete(paths):
+            count = total_size = 0
+            for path, size in paths.items():
+                try:
+                    os.remove(path)
+                    count += 1
+                    total_size += size
+                except FileNotFoundError as e:
+                    logger.error(f'could not delete stale block file {path}: {e}')
+            return count, total_size
 
-                hex_hashes = await daemon.block_hex_hashes(first, count)
-                if self.caught_up:
-                    self.logger.info('new block height {:,d} hash {}'
-                                     .format(first + count-1, hex_hashes[-1]))
-                blocks = await daemon.raw_blocks(hex_hashes)
+        if not items:
+            return
+        paths = {}
+        for item in items:
+            if isinstance(item, os.DirEntry):
+                paths[item.path] = item.stat().st_size
+            else:
+                height, size = cls.blocks.pop(item)
+                paths[cls.filename(item, height)] = size
 
-                assert count == len(blocks)
+        count, total_size = await run_in_thread(delete, paths)
+        if log:
+            logger.info(f'deleted {count:,d} stale block files, total size {total_size:,d} bytes')
 
-                # Special handling for genesis block
-                if first == 0:
-                    blocks[0] = self.coin.genesis_block(blocks[0])
-                    self.logger.info('verified genesis block with hash {}'
-                                     .format(hex_hashes[0]))
+    @classmethod
+    async def delete_blocks(cls, min_height, log):
+        blocks_to_delete = [hex_hash for hex_hash, (height, size) in cls.blocks.items()
+                            if height < min_height]
+        await cls.delete_stale(blocks_to_delete, log)
 
-                # Update our recent average block size estimate
-                size = sum(len(block) for block in blocks)
-                if count >= 10:
-                    self.ave_size = size // count
-                else:
-                    self.ave_size = (size + (10 - count) * self.ave_size) // 10
+    @classmethod
+    async def scan_files(cls):
+        # Remove stale block files
+        def scan():
+            to_delete = []
+            with os.scandir(cls.path) as it:
+                for dentry in it:
+                    if dentry.is_file():
+                        match = cls.block_regex.match(dentry.name)
+                        if match:
+                            to_delete.append(dentry)
+            return to_delete
 
-                self.blocks.extend(blocks)
-                self.cache_size += size
-                self.fetched_height += count
-                self.blocks_event.set()
+        def find_legacy_blocks():
+            with os.scandir('meta') as it:
+                return [dentry for dentry in it
+                        if dentry.is_file() and cls.legacy_del_regex.match(dentry.name)]
 
-        self.refill_event.clear()
-        return True
+        try:
+            # This only succeeds the first time with the new code
+            os.mkdir(cls.path)
+            logger.info(f'created block directory {cls.path}')
+            await cls.delete_stale(await run_in_thread(find_legacy_blocks), True)
+        except FileExistsError:
+            pass
+
+        logger.info(f'scanning block directory {cls.path}...')
+        to_delete = await run_in_thread(scan)
+        await cls.delete_stale(to_delete, True)
+
+    @classmethod
+    async def prefetch_many(cls, daemon, pairs, kind):
+        async def prefetch_one(hex_hash, height):
+            '''Read a block in chunks to a file.  As the files may not be complete they need
+            to be removed when the server starts up.'''
+            try:
+                filename = cls.filename(hex_hash, height)
+                size = await daemon.get_block(hex_hash, filename)
+                cls.blocks[hex_hash] = (height, size)
+                if kind == 'new':
+                    logger.info(f'fetched new block height {height:,d} hash {hex_hash}')
+                elif kind == 'reorg':
+                    logger.info(f'fetched reorged block height {height:,d} hash {hex_hash}')
+            except Exception as e:
+                logger.error(f'error prefetching {hex_hash}: {e}')
+            finally:
+                cls.tasks.pop(hex_hash)
+
+        # Pairs is a (height, hex_hash) iterable
+        for height, hex_hash in pairs:
+            if hex_hash not in cls.tasks and hex_hash not in cls.blocks:
+                cls.tasks[hex_hash] = await spawn(prefetch_one, hex_hash, height)
+
+    @classmethod
+    async def streamed_block(cls, hex_hash):
+        # Waits for a block to come in.
+        task = cls.tasks.get(hex_hash)
+        if task:
+            await task
+        item = cls.blocks.get(hex_hash)
+        if not item:
+            logger.error(f'block {hex_hash} missing')
+            return None
+        height, size = item
+        return cls(hex_hash, height, size)
+
+    @classmethod
+    async def stop_prefetching(cls):
+        for task in cls.tasks.values():
+            task.cancel()
+        logger.info('prefetcher stopped')
 
 
 class ChainError(Exception):
     '''Raised on error processing blocks.'''
 
 
-class BlockProcessor(object):
-    '''Process blocks and update the DB state to match.
-
-    Employ a prefetcher to prefetch blocks in batches for processing.
-    Coordinate backing up in case of chain reorganisations.
+class BlockProcessor:
+    '''Process blocks and update the DB state to match.  Prefetch blocks so they are
+    immediately available when the processor is ready for a new block.  Coordinate backing
+    up in case of chain reorganisations.
     '''
+
+    polling_delay = 5
 
     def __init__(self, env, db, daemon, notifications):
         self.env = env
         self.db = db
         self.daemon = daemon
         self.notifications = notifications
-
         self.coin = env.coin
-        self.blocks_event = asyncio.Event()
-        self.prefetcher = Prefetcher(daemon, env.coin, self.blocks_event)
-        self.logger = class_logger(__name__, self.__class__.__name__)
 
-        # Meta
-        self.next_cache_check = 0
+        self.caught_up = False
+        self.ok = True
         self.touched = set()
-        self.reorg_count = 0
-        self.height = -1
-        self.tip = None
-        self.tx_count = 0
-        self._caught_up_event = None
+        # A count >= 0 is a user-forced reorg; < 0 is a natural reorg
+        self.reorg_count = None
+        self.force_flush_arg = None
+
+        # State.  Initially taken from DB;
+        self.state = None
 
         # Caches of unflushed items.
         self.headers = []
@@ -183,115 +301,90 @@ class BlockProcessor(object):
         self.utxo_cache = {}
         self.db_deletes = []
 
-        # If the lock is successfully acquired, in-memory chain state
-        # is consistent with self.height
-        self.state_lock = asyncio.Lock()
-
-        # Signalled after backing up during a reorg
+        # Signalled after backing up during a reorg to flush session manager caches
         self.backed_up_event = asyncio.Event()
 
-    async def run_in_thread_with_lock(self, func, *args):
-        # Run in a thread to prevent blocking.  Shielded so that
-        # cancellations from shutdown don't lose work - when the task
-        # completes the data will be flushed and then we shut down.
-        # Take the state lock to be certain in-memory state is
-        # consistent and not being updated elsewhere.
-        async def run_in_thread_locked():
+        # When the lock is acquired, in-memory chain state is consistent with state.height.
+        # This is a requirement for safe flushing.
+        self.state_lock = asyncio.Lock()
+
+    async def run_with_lock(self, coro):
+        # Shielded so that cancellations from shutdown don't lose work.  Cancellation will
+        # cause fetch_and_process_blocks to block on the lock in flush(), the task completes,
+        # and then the data is flushed.  We also don't want user-signalled reorgs to happen
+        # in the middle of processing blocks; they need to wait.
+        async def run_locked():
             async with self.state_lock:
-                return await run_in_thread(func, *args)
-        return await asyncio.shield(run_in_thread_locked())
+                return await coro
+        return await asyncio.shield(run_locked())
 
-    async def check_and_advance_blocks(self, raw_blocks):
-        '''Process the list of raw blocks passed.  Detects and handles
-        reorgs.
-        '''
-        if not raw_blocks:
-            return
-        first = self.height + 1
-        blocks = [self.coin.block(raw_block, first + n)
-                  for n, raw_block in enumerate(raw_blocks)]
-        headers = [block.header for block in blocks]
-        hprevs = [self.coin.header_prevhash(h) for h in headers]
-        chain = [self.tip] + [self.coin.header_hash(h) for h in headers[:-1]]
+    async def next_block_hashes(self):
+        daemon_height = await self.daemon.height()
 
-        if hprevs == chain:
-            start = time.time()
-            await self.run_in_thread_with_lock(self.advance_blocks, blocks)
-            await self._maybe_flush()
-            if not self.db.first_sync:
-                s = '' if len(blocks) == 1 else 's'
-                blocks_size = sum(len(block) for block in raw_blocks) / 1_000_000
-                self.logger.info(f'processed {len(blocks):,d} block{s} size {blocks_size:.2f} MB '
-                                 f'in {time.time() - start:.1f}s')
-            if self._caught_up_event.is_set():
-                await self.notifications.on_block(self.touched, self.height)
-            self.touched = set()
-        elif hprevs[0] != chain[0]:
-            await self.reorg_chain()
+        first = self.state.height + 1
+        count = min(daemon_height - first + 1, self.coin.prefetch_limit(first))
+        if count:
+            hex_hashes = await self.daemon.block_hex_hashes(first, count)
+            kind = 'new' if self.caught_up else 'sync'
+            await OnDiskBlock.prefetch_many(self.daemon, enumerate(hex_hashes, start=first), kind)
         else:
-            # It is probably possible but extremely rare that what
-            # bitcoind returns doesn't form a chain because it
-            # reorg-ed the chain as it was processing the batched
-            # block hash requests.  Should this happen it's simplest
-            # just to reset the prefetcher and try again.
-            self.logger.warning('daemon blocks do not form a chain; '
-                                'resetting the prefetcher')
-            await self.prefetcher.reset_height(self.height)
+            hex_hashes = []
 
-    async def reorg_chain(self, count=None):
+        # Remove stale blocks
+        await OnDiskBlock.delete_blocks(first - 5, False)
+
+        # Return half the prefetch
+        return hex_hashes[:(count + 1) // 2], daemon_height
+
+    async def reorg_chain(self, count):
         '''Handle a chain reorganisation.
 
-        Count is the number of blocks to simulate a reorg, or None for
-        a real reorg.'''
-        if count is None:
-            self.logger.info('chain reorg detected')
+        Count is the number of blocks to simulate a reorg, or None for a real reorg.
+        This is passed in as self.reorg_count may change asynchronously.
+        '''
+        if count < 0:
+            logger.info('chain reorg detected')
         else:
-            self.logger.info(f'faking a reorg of {count:,d} blocks')
+            logger.info(f'faking a reorg of {count:,d} blocks')
         await self.flush(True)
 
-        async def get_raw_blocks(last_height, hex_hashes):
-            heights = range(last_height, last_height - len(hex_hashes), -1)
-            try:
-                blocks = [self.db.read_raw_block(height) for height in heights]
-                self.logger.info(f'read {len(blocks)} blocks from disk')
-                return blocks
-            except FileNotFoundError:
-                return await self.daemon.raw_blocks(hex_hashes)
+        start, hex_hashes = await self._reorg_hashes(count)
+        pairs = reversed(list(enumerate(hex_hashes, start=start)))
+        await OnDiskBlock.prefetch_many(self.daemon, pairs, 'reorg')
 
-        def flush_backup():
-            # self.touched can include other addresses which is
-            # harmless, but remove None.
-            self.touched.discard(None)
-            self.db.flush_backup(self.flush_data(), self.touched)
+        for hex_hash in reversed(hex_hashes):
+            if hex_hash != hash_to_hex_str(self.state.tip):
+                logger.error(f'block {hex_hash} is not tip; cannot back up')
+                return
+            block = await OnDiskBlock.streamed_block(hex_hash)
+            if not block:
+                break
+            await self.run_with_lock(run_in_thread(self.backup_block, block))
 
-        _start, last, hashes = await self.reorg_hashes(count)
-        # Reverse and convert to hex strings.
-        hashes = [hash_to_hex_str(hash) for hash in reversed(hashes)]
-        for hex_hashes in chunks(hashes, 50):
-            raw_blocks = await get_raw_blocks(last, hex_hashes)
-            await self.run_in_thread_with_lock(self.backup_blocks, raw_blocks)
-            await self.run_in_thread_with_lock(flush_backup)
-            last -= len(raw_blocks)
-        await self.prefetcher.reset_height(self.height)
+        logger.info(f'backed up to height {self.state.height:,d}')
         self.backed_up_event.set()
         self.backed_up_event.clear()
 
-    async def reorg_hashes(self, count):
-        '''Return a pair (start, last, hashes) of blocks to back up during a
+    async def _reorg_hashes(self, count):
+        '''Return a pair (start, hashes) of blocks to back up during a
         reorg.
 
         The hashes are returned in order of increasing height.  Start
         is the height of the first hash, last of the last.
         '''
-        start, count = await self.calc_reorg_range(count)
+        start, count = await self._calc_reorg_range(count)
         last = start + count - 1
-        s = '' if count == 1 else 's'
-        self.logger.info(f'chain was reorganised replacing {count:,d} '
-                         f'block{s} at heights {start:,d}-{last:,d}')
+        if count == 1:
+            logger.info(f'chain was reorganised replacing 1 block at height {start:,d}')
+        else:
+            logger.info(f'chain was reorganised replacing {count:,d} blocks at heights '
+                        f'{start:,d}-{last:,d}')
 
-        return start, last, await self.db.fs_block_hashes(start, count)
+        hashes = await self.db.fs_block_hashes(start, count)
+        hex_hashes = [hash_to_hex_str(block_hash) for block_hash in hashes]
+        return start, hex_hashes
 
-    async def calc_reorg_range(self, count):
+    async def _calc_reorg_range(self, count):
         '''Calculate the reorg range'''
 
         def diff_pos(hashes1, hashes2):
@@ -302,9 +395,10 @@ class BlockProcessor(object):
                     return n
             return len(hashes)
 
-        if count is None:
+        height = self.state.height
+        if count < 0:
             # A real reorg
-            start = self.height - 1
+            start = height - 1
             count = 1
             while start > 0:
                 hashes = await self.db.fs_block_hashes(start, count)
@@ -317,102 +411,84 @@ class BlockProcessor(object):
                 count = min(count * 2, start)
                 start -= count
 
-            count = (self.height - start) + 1
+            count = (height - start) + 1
         else:
-            start = (self.height - count) + 1
+            start = (height - count) + 1
 
         return start, count
 
-    def estimate_txs_remaining(self):
-        # Try to estimate how many txs there are to go
-        daemon_height = self.daemon.cached_height()
-        coin = self.coin
-        tail_count = daemon_height - max(self.height, coin.TX_COUNT_HEIGHT)
-        # Damp the initial enthusiasm
-        realism = max(2.0 - 0.9 * self.height / coin.TX_COUNT_HEIGHT, 1.0)
-        return (tail_count * coin.TX_PER_BLOCK +
-                max(coin.TX_COUNT - self.tx_count, 0)) * realism
-
     # - Flushing
     def flush_data(self):
-        '''The data for a flush.  The lock must be taken.'''
-        assert self.state_lock.locked()
-        return FlushData(self.height, self.tx_count, self.headers,
-                         self.tx_hashes, self.undo_infos, self.utxo_cache,
-                         self.db_deletes, self.tip)
+        '''The data for a flush.'''
+        return FlushData(self.state, self.headers, self.tx_hashes, self.undo_infos,
+                         self.utxo_cache, self.db_deletes)
 
     async def flush(self, flush_utxos):
-        def flush():
-            self.db.flush_dbs(self.flush_data(), flush_utxos,
-                              self.estimate_txs_remaining)
-        await self.run_in_thread_with_lock(flush)
+        self.force_flush_arg = None
+        # Estimate size remaining
+        daemon_height = self.daemon.cached_height()
+        tail_blocks = max(0, (daemon_height - max(self.state.height, self.coin.CHAIN_SIZE_HEIGHT)))
+        size_remaining = (max(self.coin.CHAIN_SIZE - self.state.chain_size, 0)
+                          + tail_blocks * self.coin.AVG_BLOCK_SIZE)
+        await run_in_thread(self.db.flush_dbs, self.flush_data(), flush_utxos, size_remaining)
 
-    async def _maybe_flush(self):
-        # If caught up, flush everything as client queries are
-        # performed on the DB.
-        if self._caught_up_event.is_set():
-            await self.flush(True)
-        elif time.time() > self.next_cache_check:
-            flush_arg = self.check_cache_size()
-            if flush_arg is not None:
-                await self.flush(flush_arg)
-            self.next_cache_check = time.time() + 30
-
-    def check_cache_size(self):
-        '''Flush a cache if it gets too big.'''
-        # Good average estimates based on traversal of subobjects and
-        # requesting size from Python (see deep_getsizeof).
-        one_MB = 1000*1000
-        utxo_cache_size = len(self.utxo_cache) * 205
-        db_deletes_size = len(self.db_deletes) * 57
-        hist_cache_size = self.db.history.unflushed_memsize()
-        # Roughly ntxs * 32 + nblocks * 42
-        tx_hash_size = ((self.tx_count - self.db.fs_tx_count) * 32
-                        + (self.height - self.db.fs_height) * 42)
-        utxo_MB = (db_deletes_size + utxo_cache_size) // one_MB
-        hist_MB = (hist_cache_size + tx_hash_size) // one_MB
-
-        self.logger.info('our height: {:,d} daemon: {:,d} '
-                         'UTXOs {:,d}MB hist {:,d}MB'
-                         .format(self.height, self.daemon.cached_height(),
-                                 utxo_MB, hist_MB))
-
-        # Flush history if it takes up over 20% of cache memory.
-        # Flush UTXOs once they take up 80% of cache memory.
+    async def check_cache_size_loop(self):
+        '''Signal to flush caches if they get too big.'''
+        one_MB = 1_000_000
         cache_MB = self.env.cache_MB
-        if utxo_MB + hist_MB >= cache_MB or hist_MB >= cache_MB // 5:
-            return utxo_MB >= cache_MB * 4 // 5
-        return None
+        OnDiskBlock.daemon = self.daemon
+        while True:
+            # Good average estimates based on traversal of subobjects and
+            # requesting size from Python (see deep_getsizeof).
+            utxo_cache_size = len(self.utxo_cache) * 205
+            db_deletes_size = len(self.db_deletes) * 57
+            hist_cache_size = self.db.history.unflushed_memsize()
+            # Roughly ntxs * 32 + nblocks * 42
+            tx_hash_size = ((self.state.tx_count - self.db.fs_tx_count) * 32
+                            + (self.state.height - self.db.fs_height) * 42)
+            utxo_MB = (db_deletes_size + utxo_cache_size) // one_MB
+            hist_MB = (hist_cache_size + tx_hash_size) // one_MB
 
-    def advance_blocks(self, blocks):
-        '''Synchronously advance the blocks.
+            OnDiskBlock.log_block = True
+            if hist_cache_size:
+                logger.info(f'UTXOs {utxo_MB:,d}MB hist {hist_MB:,d}MB')
 
-        It is already verified they correctly connect onto our tip.
-        '''
-        min_height = self.db.min_undo_height(self.daemon.cached_height())
-        height = self.height
-        genesis_activation = self.coin.GENESIS_ACTIVATION
+            # Flush history if it takes up over 20% of cache memory.
+            # Flush UTXOs once they take up 80% of cache memory.
+            if utxo_MB + hist_MB >= cache_MB or hist_MB >= cache_MB // 5:
+                self.force_flush_arg = utxo_MB >= cache_MB * 4 // 5
+            await sleep(30)
 
-        for block in blocks:
-            height += 1
-            is_unspendable = (is_unspendable_genesis if height >= genesis_activation
-                              else is_unspendable_legacy)
-            undo_info = self.advance_txs(block.transactions, is_unspendable)
-            if height >= min_height:
-                self.undo_infos.append((undo_info, height))
-                self.db.write_raw_block(block.raw, height)
+    async def advance_blocks(self, hex_hashes):
+        '''Process the blocks passed.  Detects and handles reorgs.'''
+        async def advance_and_maybe_flush(block):
+            await run_in_thread(self.advance_block, block)
+            if self.force_flush_arg is not None:
+                await self.flush(self.force_flush_arg)
 
-        headers = [block.header for block in blocks]
-        self.height = height
-        self.headers.extend(headers)
-        self.tip = self.coin.header_hash(headers[-1])
+        for hex_hash in hex_hashes:
+            if self.reorg_count is not None:
+                break
+            block = await OnDiskBlock.streamed_block(hex_hash)
+            if not block:
+                break
+            await self.run_with_lock(advance_and_maybe_flush(block))
 
-    def advance_txs(self, txs, is_unspendable):
-        self.tx_hashes.append(b''.join(tx_hash for tx, tx_hash in txs))
+        # If we've not caught up we have no clients for the touched set
+        if not self.caught_up:
+            self.touched = set()
+
+    def advance_block(self, block):
+        '''Advance once block.  It is already verified they correctly connect onto our tip.'''
+
+        is_unspendable = (is_unspendable_genesis if block.height >= self.coin.GENESIS_ACTIVATION
+                          else is_unspendable_legacy)
 
         # Use local vars for speed in the loops
+        state = self.state
+        tx_hashes = []
         undo_info = []
-        tx_num = self.tx_count
+        tx_num = state.tx_count
         script_hashX = self.coin.hashX_from_script
         put_utxo = self.utxo_cache.__setitem__
         spend_utxo = self.spend_utxo
@@ -420,113 +496,128 @@ class BlockProcessor(object):
         update_touched = self.touched.update
         hashXs_by_tx = []
         append_hashXs = hashXs_by_tx.append
+        append_tx_hash = tx_hashes.append
         to_le_uint32 = pack_le_uint32
         to_le_uint64 = pack_le_uint64
+        utxo_count_delta = 0
 
-        for tx, tx_hash in txs:
-            hashXs = []
-            append_hashX = hashXs.append
-            tx_numb = to_le_uint64(tx_num)[:5]
+        with block as block:
+            if self.coin.header_prevhash(block.header) != self.state.tip:
+                self.reorg_count = -1
+                return
 
-            # Spend the inputs
-            for txin in tx.inputs:
-                if txin.is_generation():
-                    continue
-                cache_value = spend_utxo(txin.prev_hash, txin.prev_idx)
-                undo_info_append(cache_value)
-                append_hashX(cache_value[:-13])
+            self.ok = False
+            for tx, tx_hash in block.iter_txs():
+                hashXs = []
+                append_hashX = hashXs.append
+                tx_numb = to_le_uint64(tx_num)[:5]
 
-            # Add the new UTXOs
-            for idx, txout in enumerate(tx.outputs):
-                # Ignore unspendable outputs
-                if is_unspendable(txout.pk_script):
-                    continue
+                # Spend the inputs
+                for txin in tx.inputs:
+                    if txin.is_generation():
+                        continue
+                    utxo_count_delta -= 1
+                    cache_value = spend_utxo(bytes(txin.prev_hash), txin.prev_idx)
+                    undo_info_append(cache_value)
+                    append_hashX(cache_value[:-13])
 
-                # Get the hashX
-                hashX = script_hashX(txout.pk_script)
-                append_hashX(hashX)
-                put_utxo(tx_hash + to_le_uint32(idx),
-                         hashX + tx_numb + to_le_uint64(txout.value))
+                # Add the new UTXOs
+                for idx, txout in enumerate(tx.outputs):
+                    # Ignore unspendable outputs
+                    if is_unspendable(txout.pk_script):
+                        continue
+                    utxo_count_delta += 1
 
-            append_hashXs(hashXs)
-            update_touched(hashXs)
-            tx_num += 1
+                    # Get the hashX
+                    hashX = script_hashX(txout.pk_script)
+                    append_hashX(hashX)
+                    put_utxo(tx_hash + to_le_uint32(idx),
+                             hashX + tx_numb + to_le_uint64(txout.value))
 
-        self.db.history.add_unflushed(hashXs_by_tx, self.tx_count)
+                append_hashXs(hashXs)
+                update_touched(hashXs)
+                append_tx_hash(tx_hash)
+                tx_num += 1
 
-        self.tx_count = tx_num
+        # Do this first - it uses the prior state
+        self.tx_hashes.append(b''.join(tx_hashes))
+        self.db.history.add_unflushed(hashXs_by_tx, state.tx_count)
         self.db.tx_counts.append(tx_num)
+        if block.height >= self.db.min_undo_height(self.daemon.cached_height()):
+            self.undo_infos.append((undo_info, block.height))
+        self.headers.append(block.header)
 
-        return undo_info
+        # Update state
+        state.height = block.height
+        state.tip = self.coin.header_hash(block.header)
+        state.chain_size += block.size
+        state.utxo_count += utxo_count_delta
+        state.tx_count = tx_num
+        self.ok = True
 
-    def backup_blocks(self, raw_blocks):
-        '''Backup the raw blocks and flush.
-
-        The blocks should be in order of decreasing height, starting at.
-        self.height.  A flush is performed once the blocks are backed up.
-        '''
+    def backup_block(self, block):
+        '''Backup the streamed block.'''
         self.db.assert_flushed(self.flush_data())
-        assert self.height >= len(raw_blocks)
+        assert block.height > 0
         genesis_activation = self.coin.GENESIS_ACTIVATION
 
-        coin = self.coin
-        for raw_block in raw_blocks:
-            # Check and update self.tip
-            block = coin.block(raw_block, self.height)
-            header_hash = coin.header_hash(block.header)
-            if header_hash != self.tip:
-                raise ChainError('backup block {} not tip {} at height {:,d}'
-                                 .format(hash_to_hex_str(header_hash),
-                                         hash_to_hex_str(self.tip),
-                                         self.height))
-            self.tip = coin.header_prevhash(block.header)
-            is_unspendable = (is_unspendable_genesis if self.height >= genesis_activation
-                              else is_unspendable_legacy)
-            self.backup_txs(block.transactions, is_unspendable)
-            self.height -= 1
-            self.db.tx_counts.pop()
+        is_unspendable = (is_unspendable_genesis if block.height >= genesis_activation
+                          else is_unspendable_legacy)
 
-        self.logger.info('backed up to height {:,d}'.format(self.height))
-
-    def backup_txs(self, txs, is_unspendable):
         # Prevout values, in order down the block (coinbase first if present)
         # undo_info is in reverse block order
-        undo_info = self.db.read_undo_info(self.height)
+        undo_info = self.db.read_undo_info(block.height)
         if undo_info is None:
-            raise ChainError('no undo information found for height {:,d}'
-                             .format(self.height))
+            raise ChainError(f'no undo information found for height {block.height:,d}')
         n = len(undo_info)
 
         # Use local vars for speed in the loops
         put_utxo = self.utxo_cache.__setitem__
         spend_utxo = self.spend_utxo
-        script_hashX = self.coin.hashX_from_script
-        touched = self.touched
+        touched_add = self.touched.add
         undo_entry_len = 13 + HASHX_LEN
 
-        for tx, tx_hash in reversed(txs):
-            for idx, txout in enumerate(tx.outputs):
-                # Spend the TX outputs.  Be careful with unspendable
-                # outputs - we didn't save those in the first place.
-                if is_unspendable(txout.pk_script):
-                    continue
+        count = 0
+        utxo_count_delta = 0
+        with block as block:
+            self.ok = False
+            for tx, tx_hash in block.iter_txs_reversed():
+                for idx, txout in enumerate(tx.outputs):
+                    # Spend the TX outputs.  Be careful with unspendable
+                    # outputs - we didn't save those in the first place.
+                    if is_unspendable(txout.pk_script):
+                        continue
 
-                # Get the hashX
-                hashX = script_hashX(txout.pk_script)
-                cache_value = spend_utxo(tx_hash, idx)
-                touched.add(cache_value[:-13])
+                    utxo_count_delta -= 1
+                    cache_value = spend_utxo(tx_hash, idx)
+                    touched_add(cache_value[:-13])
 
-            # Restore the inputs
-            for txin in reversed(tx.inputs):
-                if txin.is_generation():
-                    continue
-                n -= undo_entry_len
-                undo_item = undo_info[n:n + undo_entry_len]
-                put_utxo(txin.prev_hash + pack_le_uint32(txin.prev_idx), undo_item)
-                touched.add(undo_item[:-13])
+                # Restore the inputs
+                for txin in reversed(tx.inputs):
+                    if txin.is_generation():
+                        continue
+                    utxo_count_delta += 1
+                    n -= undo_entry_len
+                    undo_item = undo_info[n:n + undo_entry_len]
+                    put_utxo(bytes(txin.prev_hash) + pack_le_uint32(txin.prev_idx), undo_item)
+                    touched_add(undo_item[:-13])
+                count += 1
 
         assert n == 0
-        self.tx_count -= len(txs)
+
+        # Update state
+        state = self.state
+        state.height -= 1
+        state.tip = self.coin.header_prevhash(block.header)
+        state.chain_size -= block.size
+        state.utxo_count += utxo_count_delta
+        state.tx_count -= count
+
+        self.db.tx_counts.pop()
+        # self.touched can include other addresses which is harmless, but remove None.
+        self.touched.discard(None)
+        self.db.flush_backup(self.flush_data(), self.touched)
+        self.ok = True
 
     '''An in-memory UTXO cache, representing all changes to UTXO state
     since the last DB flush.
@@ -608,9 +699,9 @@ class BlockProcessor(object):
 
             if len(candidates) > 1:
                 tx_num, = unpack_le_uint64(tx_num_packed + bytes(3))
-                hash, _height = self.db.fs_tx_hash(tx_num)
-                if hash != tx_hash:
-                    assert hash is not None  # Should always be found
+                fs_hash, _height = self.db.fs_tx_hash(tx_num)
+                if fs_hash != tx_hash:
+                    assert fs_hash is not None  # Should always be found
                     continue
 
             # Key: b'u' + address_hashX + tx_idx + tx_num
@@ -623,46 +714,26 @@ class BlockProcessor(object):
                 self.db_deletes.append(udb_key)
                 return hashX + tx_num_packed + utxo_value_packed
 
-        raise ChainError('UTXO {} / {:,d} not found in "h" table'
-                         .format(hash_to_hex_str(tx_hash), tx_idx))
+        raise ChainError(f'UTXO {hash_to_hex_str(tx_hash)} / {tx_idx:,d} not found in "h" table')
 
-    async def _process_prefetched_blocks(self):
-        '''Loop forever processing blocks as they arrive.'''
-        while True:
-            if self.height == self.daemon.cached_height():
-                if not self._caught_up_event.is_set():
-                    await self._first_caught_up()
-                    self._caught_up_event.set()
-            await self.blocks_event.wait()
-            self.blocks_event.clear()
-            if self.reorg_count:
-                await self.reorg_chain(self.reorg_count)
-                self.reorg_count = 0
-            else:
-                blocks = self.prefetcher.get_prefetched_blocks()
-                await self.check_and_advance_blocks(blocks)
-
-    async def _first_caught_up(self):
-        self.logger.info(f'caught up to height {self.height}')
-        # Flush everything but with first_sync->False state.
-        first_sync = self.db.first_sync
-        self.db.first_sync = False
+    async def on_caught_up(self):
+        was_first_sync = self.state.first_sync
+        self.state.first_sync = False
         await self.flush(True)
-        if first_sync:
-            self.logger.info(f'{electrumx.version} synced to '
-                             f'height {self.height:,d}')
-        # Reopen for serving
-        await self.db.open_for_serving()
-
-    async def _first_open_dbs(self):
-        await self.db.open_for_sync()
-        self.height = self.db.db_height
-        self.tip = self.db.db_tip
-        self.tx_count = self.db.db_tx_count
+        if self.caught_up:
+            # Flush everything before notifying as client queries are performed on the DB
+            await self.notifications.on_block(self.touched, self.state.height)
+            self.touched = set()
+        else:
+            self.caught_up = True
+            if was_first_sync:
+                logger.info(f'{electrumx.version} synced to height {self.state.height:,d}')
+            # Reopen for serving
+            await self.db.open_for_serving()
 
     # --- External API
 
-    async def fetch_and_process_blocks(self, caught_up_event):
+    async def fetch_and_process_blocks(self, caught_up_event, shutdown_event):
         '''Fetch, process and index blocks from the daemon.
 
         Sets caught_up_event when first caught up.  Flushes to disk
@@ -674,166 +745,57 @@ class BlockProcessor(object):
         disk before exiting, as otherwise a significant amount of work
         could be lost.
         '''
-        self._caught_up_event = caught_up_event
-        await self._first_open_dbs()
-        try:
-            async with TaskGroup() as group:
-                await group.spawn(self.prefetcher.main_loop(self.height))
-                await group.spawn(self._process_prefetched_blocks())
-        # Don't flush for arbitrary exceptions as they might be a cause or consequence of
-        # corrupted data
-        except CancelledError:
-            self.logger.info('flushing to DB for a clean shutdown...')
+        self.state = OnDiskBlock.state = (await self.db.open_for_sync()).copy()
+        await OnDiskBlock.scan_files()
+
+        while True:
+            try:
+                show_summary = True
+                while True:
+                    hex_hashes, daemon_height = await self.next_block_hashes()
+                    if show_summary:
+                        show_summary = False
+                        behind = daemon_height - self.state.height
+                        if behind > 0:
+                            logger.info(f'catching up to daemon height {daemon_height:,d} '
+                                        f'({behind:,d} blocks behind)')
+                        else:
+                            logger.info(f'caught up to daemon height {daemon_height:,d}')
+
+                    if hex_hashes:
+                        await self.advance_blocks(hex_hashes)
+                    else:
+                        await self.on_caught_up()
+                        caught_up_event.set()
+                        await sleep(self.polling_delay)
+
+                    if self.reorg_count is not None:
+                        await self.reorg_chain(self.reorg_count)
+                        self.reorg_count = None
+                        show_summary = True
+                # Don't flush for arbitrary exceptions as they might be a cause or
+                # consequence of corrupted data
+            except CancelledError:
+                if shutdown_event.is_set():
+                    await OnDiskBlock.stop_prefetching()
+                    await self.run_with_lock(self.flush_if_safe())
+                    return
+                logger.exception('ignoring unexpected cancellation')
+
+    async def flush_if_safe(self):
+        if self.ok:
+            logger.info('flushing to DB for a clean shutdown...')
             await self.flush(True)
+            logger.info('flushed cleanly')
+        else:
+            logger.warning('not flushing to DB as data in memory is incomplete')
 
     def force_chain_reorg(self, count):
-        '''Force a reorg of the given number of blocks.
-
-        Returns True if a reorg is queued, false if not caught up.
+        '''Force a reorg of the given number of blocks.  Returns True if a reorg is queued.
+        During initial sync we don't store undo information so cannot fake a reorg until
+        caught up.
         '''
-        if self._caught_up_event.is_set():
+        if self.caught_up:
             self.reorg_count = count
-            self.blocks_event.set()
             return True
         return False
-
-
-class DecredBlockProcessor(BlockProcessor):
-    async def calc_reorg_range(self, count):
-        start, count = await super().calc_reorg_range(count)
-        if start > 0:
-            # A reorg in Decred can invalidate the previous block
-            start -= 1
-            count += 1
-        return start, count
-
-
-class NameIndexBlockProcessor(BlockProcessor):
-
-    def advance_txs(self, txs, is_unspendable):
-        result = super().advance_txs(txs, is_unspendable)
-
-        tx_num = self.tx_count - len(txs)
-        script_name_hashX = self.coin.name_hashX_from_script
-        update_touched = self.touched.update
-        hashXs_by_tx = []
-        append_hashXs = hashXs_by_tx.append
-
-        for tx, _tx_hash in txs:
-            hashXs = []
-            append_hashX = hashXs.append
-
-            # Add the new UTXOs and associate them with the name script
-            for txout in tx.outputs:
-                # Get the hashX of the name script.  Ignore non-name scripts.
-                hashX = script_name_hashX(txout.pk_script)
-                if hashX:
-                    append_hashX(hashX)
-
-            append_hashXs(hashXs)
-            update_touched(hashXs)
-            tx_num += 1
-
-        self.db.history.add_unflushed(hashXs_by_tx, self.tx_count - len(txs))
-
-        return result
-
-
-class LTORBlockProcessor(BlockProcessor):
-
-    def advance_txs(self, txs, is_unspendable):
-        self.tx_hashes.append(b''.join(tx_hash for tx, tx_hash in txs))
-
-        # Use local vars for speed in the loops
-        undo_info = []
-        tx_num = self.tx_count
-        script_hashX = self.coin.hashX_from_script
-        put_utxo = self.utxo_cache.__setitem__
-        spend_utxo = self.spend_utxo
-        undo_info_append = undo_info.append
-        update_touched = self.touched.update
-        to_le_uint32 = pack_le_uint32
-        to_le_uint64 = pack_le_uint64
-
-        hashXs_by_tx = [set() for _ in txs]
-
-        # Add the new UTXOs
-        for (tx, tx_hash), hashXs in zip(txs, hashXs_by_tx):
-            add_hashXs = hashXs.add
-            tx_numb = to_le_uint64(tx_num)[:5]
-
-            for idx, txout in enumerate(tx.outputs):
-                # Ignore unspendable outputs
-                if is_unspendable(txout.pk_script):
-                    continue
-
-                # Get the hashX
-                hashX = script_hashX(txout.pk_script)
-                add_hashXs(hashX)
-                put_utxo(tx_hash + to_le_uint32(idx),
-                         hashX + tx_numb + to_le_uint64(txout.value))
-            tx_num += 1
-
-        # Spend the inputs
-        # A separate for-loop here allows any tx ordering in block.
-        for (tx, tx_hash), hashXs in zip(txs, hashXs_by_tx):
-            add_hashXs = hashXs.add
-            for txin in tx.inputs:
-                if txin.is_generation():
-                    continue
-                cache_value = spend_utxo(txin.prev_hash, txin.prev_idx)
-                undo_info_append(cache_value)
-                add_hashXs(cache_value[:-13])
-
-        # Update touched set for notifications
-        for hashXs in hashXs_by_tx:
-            update_touched(hashXs)
-
-        self.db.history.add_unflushed(hashXs_by_tx, self.tx_count)
-
-        self.tx_count = tx_num
-        self.db.tx_counts.append(tx_num)
-
-        return undo_info
-
-    def backup_txs(self, txs, is_unspendable):
-        undo_info = self.db.read_undo_info(self.height)
-        if undo_info is None:
-            raise ChainError('no undo information found for height {:,d}'
-                             .format(self.height))
-
-        # Use local vars for speed in the loops
-        put_utxo = self.utxo_cache.__setitem__
-        spend_utxo = self.spend_utxo
-        script_hashX = self.coin.hashX_from_script
-        add_touched = self.touched.add
-        undo_entry_len = 13 + HASHX_LEN
-
-        # Restore coins that had been spent
-        # (may include coins made then spent in this block)
-        n = 0
-        for tx, tx_hash in txs:
-            for txin in tx.inputs:
-                if txin.is_generation():
-                    continue
-                undo_item = undo_info[n:n + undo_entry_len]
-                put_utxo(txin.prev_hash + pack_le_uint32(txin.prev_idx), undo_item)
-                add_touched(undo_item[:-13])
-                n += undo_entry_len
-
-        assert n == len(undo_info)
-
-        # Remove tx outputs made in this block, by spending them.
-        for tx, tx_hash in txs:
-            for idx, txout in enumerate(tx.outputs):
-                # Spend the TX outputs.  Be careful with unspendable
-                # outputs - we didn't save those in the first place.
-                if is_unspendable(txout.pk_script):
-                    continue
-
-                # Get the hashX
-                hashX = script_hashX(txout.script)
-                cache_value = spend_utxo(tx_hash, idx)
-                add_touched(cache_value[:-13])
-
-        self.tx_count -= len(txs)
