@@ -1,4 +1,4 @@
-version = 'ElectrumX 1.15.0'
+version = 'ElectrumX 1.20.2'
 version_short = version.split()[-1]
 
 from electrumx.server.controller import Controller
