@@ -1,28 +1,9 @@
-# Copyright (c) 2016-2017, Neil Booth
+# Copyright (c) 2016-2021, Neil Booth
 # Copyright (c) 2017, the ElectrumX authors
 #
 # All rights reserved.
 #
-# The MIT License (MIT)
-#
-# Permission is hereby granted, free of charge, to any person obtaining
-# a copy of this software and associated documentation files (the
-# "Software"), to deal in the Software without restriction, including
-# without limitation the rights to use, copy, modify, merge, publish,
-# distribute, sublicense, and/or sell copies of the Software, and to
-# permit persons to whom the Software is furnished to do so, subject to
-# the following conditions:
-#
-# The above copyright notice and this permission notice shall be
-# included in all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+# This file is licensed under the Open BSV License version 3, see LICENCE for details.
 
 '''Module providing coin abstraction.
 
@@ -30,23 +11,15 @@ Anything coin-specific should go in this file and be subclassed where
 necessary for appropriate handling.
 '''
 
-from collections import namedtuple
 import re
-import struct
 from decimal import Decimal
 from hashlib import sha256
 
-import electrumx.lib.util as util
-from electrumx.lib.hash import Base58, hash160, double_sha256, hash_to_hex_str
+from electrumx.lib import util
+from electrumx.lib.hash import Base58, double_sha256
 from electrumx.lib.hash import HASHX_LEN
 from electrumx.lib.script import ScriptPubKey
-import electrumx.lib.tx as lib_tx
-import electrumx.server.block_processor as block_proc
-import electrumx.server.daemon as daemon
 from electrumx.server.session import ElectrumX
-
-
-Block = namedtuple("Block", "raw header transactions")
 
 
 class CoinError(Exception):
@@ -62,25 +35,11 @@ class Coin:
     # Not sure if these are coin-specific
     RPC_URL_REGEX = re.compile('.+@(\\[[0-9a-fA-F:]+\\]|[^:]+)(:[0-9]+)?')
     VALUE_PER_COIN = 100000000
-    CHUNK_SIZE = 2016
-    BASIC_HEADER_SIZE = 80
-    STATIC_BLOCK_HEADERS = True
     SESSIONCLS = ElectrumX
     DEFAULT_MAX_SEND = 1000000
-    DESERIALIZER = lib_tx.Deserializer
-    DAEMON = daemon.Daemon
-    BLOCK_PROCESSOR = block_proc.BlockProcessor
-    HEADER_VALUES = ('version', 'prev_block_hash', 'merkle_root', 'timestamp',
-                     'bits', 'nonce')
-    HEADER_UNPACK = struct.Struct('< I 32s 32s I I I').unpack_from
     P2PKH_VERBYTE = bytes.fromhex("00")
     P2SH_VERBYTES = [bytes.fromhex("05")]
-    XPUB_VERBYTES = bytes.fromhex("0488b21e")
-    XPRV_VERBYTES = bytes.fromhex("0488ade4")
     RPC_PORT = 8332
-    WIF_BYTE = bytes.fromhex("80")
-    ENCODE_CHECK = Base58.encode_check
-    DECODE_CHECK = Base58.decode_check
     GENESIS_HASH = ('000000000019d6689c085ae165831e93'
                     '4ff763ae46a2a6c172b3f1b60a8ce26f')
     GENESIS_ACTIVATION = 100_000_000
@@ -93,10 +52,9 @@ class Coin:
         '''Return a coin class given name and network.
 
         Raise an exception if unrecognised.'''
-        req_attrs = ['TX_COUNT', 'TX_COUNT_HEIGHT', 'TX_PER_BLOCK']
+        req_attrs = ['CHAIN_SIZE', 'CHAIN_SIZE_HEIGHT', 'AVG_BLOCK_SIZE']
         for coin in util.subclasses(Coin):
-            if (coin.NAME.lower() == name.lower() and
-                    coin.NET.lower() == net.lower()):
+            if coin.NAME.lower() == name.lower() and coin.NET.lower() == net.lower():
                 coin_req_attrs = req_attrs.copy()
                 missing = [attr for attr in coin_req_attrs
                            if not hasattr(coin, attr)]
@@ -121,61 +79,14 @@ class Coin:
         return url + '/'
 
     @classmethod
-    def max_fetch_blocks(cls, height):
-        if height < 130000:
-            return 1000
-        return 100
-
-    @classmethod
-    def genesis_block(cls, block):
-        '''Check the Genesis block is the right one for this coin.
-
-        Return the block less its unspendable coinbase.
-        '''
-        header = cls.block_header(block, 0)
-        header_hex_hash = hash_to_hex_str(cls.header_hash(header))
-        if header_hex_hash != cls.GENESIS_HASH:
-            raise CoinError('genesis block has hash {} expected {}'
-                            .format(header_hex_hash, cls.GENESIS_HASH))
-
-        return header + bytes(1)
-
-    @classmethod
     def hashX_from_script(cls, script):
         '''Returns a hashX from a script.'''
         return sha256(script).digest()[:HASHX_LEN]
-
-    @staticmethod
-    def lookup_xverbytes(verbytes):
-        '''Return a (is_xpub, coin_class) pair given xpub/xprv verbytes.'''
-        for coin in util.subclasses(Coin):
-            if verbytes == coin.XPUB_VERBYTES:
-                return True, coin
-            if verbytes == coin.XPRV_VERBYTES:
-                return False, coin
-        raise CoinError('version bytes unrecognised')
 
     @classmethod
     def address_to_hashX(cls, address):
         '''Return a hashX given a coin address.'''
         return cls.hashX_from_script(cls.pay_to_address_script(address))
-
-    @classmethod
-    def P2PKH_address_from_hash160(cls, hash160):
-        '''Return a P2PKH address given a public key.'''
-        assert len(hash160) == 20
-        return cls.ENCODE_CHECK(cls.P2PKH_VERBYTE + hash160)
-
-    @classmethod
-    def P2PKH_address_from_pubkey(cls, pubkey):
-        '''Return a coin address given a public key.'''
-        return cls.P2PKH_address_from_hash160(hash160(pubkey))
-
-    @classmethod
-    def P2SH_address_from_hash160(cls, hash160):
-        '''Return a coin address given a hash160.'''
-        assert len(hash160) == 20
-        return cls.ENCODE_CHECK(cls.P2SH_VERBYTES[0] + hash160)
 
     @classmethod
     def hash160_to_P2PKH_script(cls, hash160):
@@ -191,7 +102,7 @@ class Coin:
 
         Pass the address (either P2PKH or P2SH) in base58 form.
         '''
-        raw = cls.DECODE_CHECK(address)
+        raw = Base58.decode_check(address)
 
         # Require version byte(s) plus hash160.
         verbyte = -1
@@ -207,14 +118,6 @@ class Coin:
         raise CoinError('invalid address: {}'.format(address))
 
     @classmethod
-    def privkey_WIF(cls, privkey_bytes, compressed):
-        '''Return the private key encoded in Wallet Import Format.'''
-        payload = bytearray(cls.WIF_BYTE) + privkey_bytes
-        if compressed:
-            payload.append(0x01)
-        return cls.ENCODE_CHECK(payload)
-
-    @classmethod
     def header_hash(cls, header):
         '''Given a header return hash'''
         return double_sha256(header)
@@ -223,33 +126,6 @@ class Coin:
     def header_prevhash(cls, header):
         '''Given a header return previous hash'''
         return header[4:36]
-
-    @classmethod
-    def static_header_offset(cls, height):
-        '''Given a header height return its offset in the headers file.
-
-        If header sizes change at some point, this is the only code
-        that needs updating.'''
-        assert cls.STATIC_BLOCK_HEADERS
-        return height * cls.BASIC_HEADER_SIZE
-
-    @classmethod
-    def static_header_len(cls, height):
-        '''Given a header height return its length.'''
-        return (cls.static_header_offset(height + 1)
-                - cls.static_header_offset(height))
-
-    @classmethod
-    def block_header(cls, block, height):
-        '''Returns the block header given a block and its height.'''
-        return block[:cls.static_header_len(height)]
-
-    @classmethod
-    def block(cls, raw_block, height):
-        '''Return a Block namedtuple given a raw block and its height.'''
-        header = cls.block_header(raw_block, height)
-        txs = cls.DESERIALIZER(raw_block, start=len(header)).read_tx_block()
-        return Block(raw_block, header, txs)
 
     @classmethod
     def decimal_value(cls, value):
@@ -261,20 +137,22 @@ class Coin:
         return Decimal(value) / cls.VALUE_PER_COIN
 
     @classmethod
-    def warn_old_client_on_tx_broadcast(cls, _client_ver):
-        return False
+    def prefetch_limit(cls, height):
+        if height <= 650_000:
+            return 100
+        return 10
 
 
 class BitcoinSV(Coin):
     NAME = "BitcoinSV"
-    TX_COUNT = 267318795
-    TX_COUNT_HEIGHT = 557037
-    TX_PER_BLOCK = 400
+    CHAIN_SIZE = 7_809_061_081_045
+    CHAIN_SIZE_HEIGHT = 761_539
+    AVG_BLOCK_SIZE = 800_000_000
     PEERS = [
-        'electrumx.bitcoinsv.io s',
         'satoshi.vision.cash s',
         'sv.usebsv.com s t',
         'sv.satoshi.io s t',
+        'sv2.satoshi.io s t',
     ]
     GENESIS_ACTIVATION = 620_538
 
@@ -282,17 +160,16 @@ class BitcoinSV(Coin):
 class BitcoinTestnetMixin:
     SHORTNAME = "XTN"
     NET = "testnet"
-    XPUB_VERBYTES = bytes.fromhex("043587cf")
-    XPRV_VERBYTES = bytes.fromhex("04358394")
     P2PKH_VERBYTE = bytes.fromhex("6f")
     P2SH_VERBYTES = [bytes.fromhex("c4")]
     WIF_BYTE = bytes.fromhex("ef")
     GENESIS_HASH = ('000000000933ea01ad0ee984209779ba'
                     'aec3ced90fa3f408719526f8d77f4943')
     REORG_LIMIT = 8000
-    TX_COUNT = 12242438
-    TX_COUNT_HEIGHT = 1035428
-    TX_PER_BLOCK = 21
+    CHAIN_SIZE = 26_584_216_544
+    CHAIN_SIZE_HEIGHT = 1_454_438
+    AVG_BLOCK_SIZE = 200_000
+
     RPC_PORT = 18332
     PEER_DEFAULT_PORTS = {'t': '51001', 's': '51002'}
 
@@ -311,16 +188,14 @@ class BitcoinSVScalingTestnet(BitcoinSVTestnet):
     PEERS = [
         'stn-server.electrumsv.io t51001 s51002',
     ]
-    TX_COUNT = 2015
-    TX_COUNT_HEIGHT = 5711
-    TX_PER_BLOCK = 5000
+    CHAIN_SIZE = 20_000
+    CHAIN_SIZE_HEIGHT = 100
+    AVG_BLOCK_SIZE = 2_000_000_000
     GENESIS_ACTIVATION = 14_896
 
     @classmethod
-    def max_fetch_blocks(cls, height):
-        if height <= 10:
-            return 100
-        return 3
+    def prefetch_limit(cls, height):
+        return 8
 
 
 class BitcoinSVRegtest(BitcoinSVTestnet):
@@ -328,8 +203,9 @@ class BitcoinSVRegtest(BitcoinSVTestnet):
     GENESIS_HASH = ('0f9188f13cb7b2c71f2a335e3a4fc328'
                     'bf5beb436012afca590b1a11466e2206')
     PEERS = []
-    TX_COUNT = 1
-    TX_COUNT_HEIGHT = 1
+    CHAIN_SIZE = 20_000
+    CHAIN_SIZE_HEIGHT = 100
+    AVG_BLOCK_SIZE = 1_000_000
     GENESIS_ACTIVATION = 10_000
 
 
