@@ -1,9 +1,8 @@
-# Copyright (c) 2017, Neil Booth
+# Copyright (c) 2017-2021, Neil Booth
 #
 # All rights reserved.
 #
-# See the file "LICENCE" for information about the copyright
-# and warranty status of this software.
+# This file is licensed under the Open BSV License version 3, see LICENCE for details.
 
 '''Class for server environment configuration and defaults.'''
 
@@ -49,7 +48,7 @@ class EnvBase(object):
             return int(value)
         except Exception:
             raise cls.Error('cannot convert envvar {} value {} to an integer'
-                            .format(envvar, value))
+                            .format(envvar, value)) from None
 
     @classmethod
     def custom(cls, envvar, default, parse):
