@@ -1,22 +1,21 @@
-# Copyright (c) 2016-2018, Neil Booth
+# Copyright (c) 2016-2021, Neil Booth
 #
 # All rights reserved.
 #
-# See the file "LICENCE" for information about the copyright
-# and warranty status of this software.
+# This file is licensed under the Open BSV License version 3, see LICENCE for details.
 
 '''Mempool handling.'''
 
 import itertools
 import time
 from abc import ABC, abstractmethod
-from asyncio import Lock
 from collections import defaultdict
 
 import attr
 from aiorpcx import TaskGroup, run_in_thread, sleep
 
 from electrumx.lib.hash import hash_to_hex_str, hex_str_to_hash
+from electrumx.lib.tx import read_tx
 from electrumx.lib.util import class_logger, chunks
 from electrumx.server.db import UTXO
 
@@ -113,15 +112,24 @@ class MemPool(object):
 
     async def _logging(self, synchronized_event):
         '''Print regular logs of mempool stats.'''
+        def fmt_amt(amount):
+            whole, frac = divmod(amount, 100_000_000)
+            frac = f'{frac:08}'.rstrip('0') or '0'
+            frac = f'{frac:8}'
+            return f'{whole:,d}.{frac} BSV'
+
         self.logger.info('beginning processing of daemon mempool.  '
                          'This can take some time...')
-        start = time.time()
+        start = time.monotonic()
         await synchronized_event.wait()
-        elapsed = time.time() - start
+        elapsed = time.monotonic() - start
         self.logger.info(f'synced in {elapsed:.2f}s')
         while True:
-            mempool_size = sum(tx.size for tx in self.txs.values()) / 1_000_000
-            self.logger.info(f'{len(self.txs):,d} txs {mempool_size:.2f} MB '
+            mempool_size = sum(tx.size for tx in self.txs.values())
+            fees = sum(tx.fee for tx in self.txs.values())
+            sats_byte = fees / (mempool_size or 1)
+            self.logger.info(f'{len(self.txs):,d} txs {mempool_size / 1_000_000:.2f} MB '
+                             f'fees {fmt_amt(fees)} ({sats_byte:.3f} sats/b) '
                              f'touching {len(self.hashXs):,d} addresses')
             await sleep(self.log_status_secs)
             await synchronized_event.wait()
@@ -139,7 +147,7 @@ class MemPool(object):
         deferred = {}
         unspent = set(utxo_map)
         # Try to find all prevouts so we can accept the TX
-        for hash, tx in tx_map.items():
+        for tx_hash, tx in tx_map.items():
             in_pairs = []
             try:
                 for prevout in tx.prevouts:
@@ -150,7 +158,7 @@ class MemPool(object):
                         utxo = txs[prev_hash].out_pairs[prev_index]
                     in_pairs.append(utxo)
             except KeyError:
-                deferred[hash] = tx
+                deferred[tx_hash] = tx
                 continue
 
             # Spend the prevouts
@@ -160,13 +168,13 @@ class MemPool(object):
             tx.in_pairs = tuple(in_pairs)
             # Avoid negative fees if dealing with generation-like transactions
             # because some in_parts would be missing
-            tx.fee = max(0, (sum(v for _, v in tx.in_pairs) -
-                             sum(v for _, v in tx.out_pairs)))
-            txs[hash] = tx
+            tx.fee = max(0, (sum(v for _, v in tx.in_pairs)
+                             - sum(v for _, v in tx.out_pairs)))
+            txs[tx_hash] = tx
 
             for hashX, _value in itertools.chain(tx.in_pairs, tx.out_pairs):
                 touched.add(hashX)
-                hashXs[hashX].add(hash)
+                hashXs[hashX].add(tx_hash)
 
         return deferred, {prevout: utxo_map[prevout] for prevout in unspent}
 
@@ -220,8 +228,6 @@ class MemPool(object):
             for hashes in chunks(new_hashes, 200):
                 coro = self._fetch_and_accept(hashes, all_hashes, touched)
                 await group.spawn(coro)
-            if mempool_height != self.api.db_height():
-                raise DBSyncError
 
             tx_map = {}
             utxo_map = {}
@@ -248,15 +254,15 @@ class MemPool(object):
 
         def deserialize_txs():    # This function is pure
             to_hashX = self.coin.hashX_from_script
-            deserializer = self.coin.DESERIALIZER
+            read_tx_and_size = read_tx
 
             txs = {}
-            for hash, raw_tx in zip(hashes, raw_txs):
+            for tx_hash, raw_tx in zip(hashes, raw_txs):
                 # The daemon may have evicted the tx from its
                 # mempool or it may have gotten in a block
                 if not raw_tx:
                     continue
-                tx, tx_size = deserializer(raw_tx).read_tx_and_vsize()
+                tx, tx_size = read_tx_and_size(raw_tx, 0)
                 # Convert the inputs and outputs into (hashX, value) pairs
                 # Drop generation-like inputs from MemPoolTx.prevouts
                 txin_pairs = tuple((txin.prev_hash, txin.prev_idx)
@@ -264,8 +270,8 @@ class MemPool(object):
                                    if not txin.is_generation())
                 txout_pairs = tuple((to_hashX(txout.pk_script), txout.value)
                                     for txout in tx.outputs)
-                txs[hash] = MemPoolTx(txin_pairs, None, txout_pairs,
-                                      0, tx_size)
+                txs[tx_hash] = MemPoolTx(txin_pairs, None, txout_pairs,
+                                         0, tx_size)
             return txs
 
         # Thread this potentially slow operation so as not to block
@@ -294,6 +300,10 @@ class MemPool(object):
             await group.spawn(self._refresh_hashes(synchronized_event))
             await group.spawn(self._logging(synchronized_event))
 
+            async for task in group:
+                if not task.cancelled():
+                    task.result()
+
     async def balance_delta(self, hashX):
         '''Return the unconfirmed amount in the mempool for hashX.
 
@@ -301,8 +311,8 @@ class MemPool(object):
         '''
         value = 0
         if hashX in self.hashXs:
-            for hash in self.hashXs[hashX]:
-                tx = self.txs[hash]
+            for hash_ in self.hashXs[hashX]:
+                tx = self.txs[hash_]
                 value -= sum(v for h168, v in tx.in_pairs if h168 == hashX)
                 value += sum(v for h168, v in tx.out_pairs if h168 == hashX)
         return value
